requirements.txt
Flask
mysql-connector-python==8.0.24
mysql-connector